+++ backend/fittoring/src/main/java/fittoring/application/member/service/MemberService.java
@@ -4,17 +4,15 @@ import fittoring.application.exception.BusinessErrorMessage;
 import fittoring.application.exception.ForbiddenException;
 import fittoring.application.exception.NotFoundMemberException;
 import fittoring.application.image.service.ImageService;
+import fittoring.application.member.presentation.dto.response.MyInfoResponse;
+import fittoring.application.member.presentation.dto.response.MyInfoSummaryResponse;
+import fittoring.application.member.repository.MemberRepository;
+import fittoring.application.mentoring.repository.MentoringRepository;
 import fittoring.domain.model.Image;
 import fittoring.domain.model.ImageType;
 import fittoring.domain.model.Member;
 import fittoring.domain.model.MemberRole;
 import fittoring.domain.model.Mentoring;
-import fittoring.application.member.repository.MemberRepository;
-import fittoring.application.mentoring.repository.MentoringRepository;
-import fittoring.admin.presentation.dto.AdminMemberResponse;
-import fittoring.application.member.presentation.dto.response.MyInfoResponse;
-import fittoring.application.member.presentation.dto.response.MyInfoSummaryResponse;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -65,15 +63,4 @@ public class MemberService {
         return true;
     }
 
-    public List<AdminMemberResponse> findAllForAdmin(Long memberId) {
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new NotFoundMemberException(BusinessErrorMessage.MEMBER_NOT_FOUND.getMessage()));
-        if (MemberRole.isNotAdmin(member.getRole())) {
-            throw new ForbiddenException(BusinessErrorMessage.FORBIDDEN_MEMBER.getMessage());
-        }
-        List<Member> members = memberRepository.findAllByOrderByRoleAsc();
-        return members.stream()
-                .map(AdminMemberResponse::from)
-                .toList();
-    }
 }
